Quickly and incompletely fixes problem mentioned in #57; pls review
I now delete all properties whose names start with an underscore from LevelService.getData() whenever the close button on the hierarchy modal is clicked.
This is not **done** when the modal is closed via ESC. There should be a better place for the removal code.

The current solution has the highly undesirable side-effect that all view state information (collapsed nodes, rotation) is lost when the modal is closed and re-opened.