now handles sub-graphs gracefully for drawing and collapsing
Sub-graphs here refers to parts of the hierarchy that are not connected to the
topmost level. Such parts have not been taken into account when calculating
_visible attributes for individual nodes. That caused them to be incollapsible.
It also caused some serious trouble that led to items not being drawn at all.

Fixes #59 and #87.

Also did quite some cleaning in emuhierarchy.js and HierarchyLayoutService.js