addresses #101
The nodes and links are now resized by the inverse of their containing SVG's
zoom factor. This stops them from changing their size while the user zooms.

This commit does indeed fix the issue but I'm not sure it's the best way to do
it. Zooming the whole SVG and then re-resizing most of its children to counter
the effect of zooming just doesn't feel right.

Also, there are several issues left:

(a) Really large bundles might still be unviewable because the possible zoom
range is limited. The limit is set by emuhierarchy.js's scope.scaleExtent
variable (line 34). If we set it unreasonably high to avoid 'unviewable
bundles', the empty spaces expand just as unreasonably. We would run the risk
that users of smaller bundles (which form the majority) zoom somewhere into the
nirvana.

(b) The resizing calculation is not done immediately upon zooming, but rather
when the hierarchy is completely re-rendered. In the meantime, the whole thing
looks pretty strange. The user can trigger a re-render by clicking a node (thus
opening its context menu), or re-opening the hierarchy view, or double-rotating.
We can also trigger it programmatically, by calling scope.render() from within
scope.zoom(). This is done but commented out on line 167. The problem is that
zooming might get laggy, especially on older machines.