fixes #110 by wiping all svg children on every render()
the problem is that when a new bundle is loaded, node and link IDs might be
duplicated thus causing some links from the new bundle not to be drawn.

This should be addressed by cleaning the SVG when loading a bundle, but for now
it's easier to clean it on every render(). As an undesired side effect, I have
to disable all graphical node and link transitions.