bugfix: editing labels in hierarchy view sometimes saved wrong text
This bug was unreported and concerns opening the context menu of a node in
hierarchy view, *not* changing the label input and then "saving" by pressing
return.

There are two cases:

If you had edited a label in hierarchy view before, that previously edited text
would also be saved to the new node.

If you had not edited a label in hierarchy view before, the data would remain
alright but this "null action" would be added to the history stack.