destroy modals while the're invisible
this fixes #106.

Up until now, modals were not destroyed when closed. That means e.g. that you
can open a modal, change some input field, close and re-open the modal (without
saving), et voila – your changes are still there. In the case of bug #106 this
means that watches are active in the background and can do all kinds of
cruelsome data-invalidating stuff or kill your cat while you're not looking.

Modals were only destroyed when replaced by other modals.

Now, one of the inner containers contained in the <modal> has an ng-if so it is
deleted safely from the DOM. The <modal> itself stays intact and can still be
smoothly and eye-candily transitioned into the fourth dimension (or was it the
eighth? I'm sure it was some power of two).