moved the logic for limiting scrolling/panning
Before, it was in the function getOrientatedTransform. It didn't really belong
there, plus that function is called from within render() which causes trouble.
The logic is now in a function of its own (limitPanning()) and that is called
from within zoom(), wihch seems to be a more appropriate way to call it.