bug-fix and start of re-factor viewState.hierarchyState
Fixes a yet undocumented bug where sometimes all shortcuts in hierarchy view
would fail, throwing an error like "viewState.hierarchyState.getInputFocus is
not a function". The problem was that after some actions (eg disconnect/
reconnect) the viewState was re-initialized but that worked only on the
variables. The functions in viewState.hierarchyState would just be thrown away.

Now there are some more functions viewState.*hierarchy*() that need to be moved
into the viewState.hierarchyState object.