largely reconsidered how zooming and panning work
Zooming on the hierarchy has lately grown into a very compex thing. Basically,
the SVG was scaled according to a zoom factor and then most elements were scaled
back according to that factor's inverse. This very not neat behaviour was due to
the code's evolution.

In my new approach, zooming no longer involves any scale() transforms. This is
because in the hierarchy view, zooming doesn't really enlarge anything. It
rather sets nodes farther apart. This is not what "zooming vector graphics"
means.

It is now also possible (albeit in the code, not yet in the webapp config) to
configure whether zooming and/or panning should be enabled on what I call the
cross axis (ie, not the time axis).

fixes #150
addresses #151 (by making it easy to add panning and zooming on the cross axis)