replaced $http.get().success and .error by .then()
The methods $http(...).success() and .error() were promise handling functions
only available in $http's promises. They have been deprecated for a while and
were removed in Angular 1.6.0.

This patch replaces them as suggested in:
https://github.com/angular/angular.js/commit/b54a39e2029005e0572fbd2ac0e8f6a4e5d69014

NB: The referenced document offers two alternatives that are slightly different
with regards to error handling. I chose the one that retains the behavior we had
before, although the other one might be preferable.