LevelService and LinkService create hash maps but leave caching to others
Previously, a small number of get*() functions on the two services stored their result in a hash map, so they would not need re-calculating when called with the same parameters.
However, this gave rise to the problem of cache invalidation.
They now offer a number of getHashMap*() functions that create hash maps based on the current state, but they do not store these maps.