bugfix: live updates of formants being corrected
While the formant correction tool is being used, it draws the old and
the new value of the formant track. This was, however, not visible for
the user, because the drawMarkup() routine almost instantly clears these
drawings away. Sometimes you could see what was drawn as a brief flash.

I now moved this code from the mousemove handler to drawMarkup. It works
now.