fixed a number of subtle math errors related to Start_Time in SSSF data
These were discovered with databases that have Start_Time values of about
20-30 milliseconds and where the formant correction tool was used.

The correction tool would change samples of the formant track that were
not directly under the mouse cursor, but rather a few milliseconds to
the right. This should now work.

Moreover, the selection of which SSFF samples are in the current
viewport was not quite right. This resulted in some samples on the left
side of the viewport simply not being displayed.

The math was correct when Start_Time was 0, and nearly right when
Start_Time was nearly 0, so only some databases were affected.
Particularly those that had imported signal processing results from
Praat.