fix snap boundary to top/bottom
For the snap boundary feature (invoked via shortcut b or t), the
neighbor level was determined wrongly: By searching the current level
in the bundle’s annotation data and using the neighbor in that array.

Now, we look at the current perspective and determine the neighbor level
from there. This is what users see as the level order when invoking the
feature.