allow fileFormat for ssff tracks in config schema
This matches the feature we've been testing since emuR 2.4.2, where
tracks can be stored in Rda instead of SSFF format.

The present commit does not attempt to render Rda files in the browser,
and it is not clear whether we will support that in the future. The
present commit only allows databases using the feature to still be
loaded in the EMU-webApp. Up until now, serve() failed for these
databases because the EMU-webApp spotted a schema violation.