introduce query parameter: disableBundleListSidebar
When you call the EMU-webApp, e.g. via an iframe, and append
?disableBundleListSidebar to the URL, the bundle list sidebar will not
be open by default and the user will not be able to open it. The
respective button in the top row will be invisible and the keyboard
shortcut will also be disabled.

It does not matter if the query parameter gets any value, e.g.
?disableBundleListSidebar=true. It is only relevant that the parameter
is present.

This is a feature request by @julianpoemp, who also provided an initial
implementation.