In EMBEDDED mode, publish all changes to parent window without saving
The user continues to have a save button, but all changes will also be
published to the parent window as soon as they are made (not waiting for
the user to hit the save button).

The object that is published via window.parent.postMessage() gets a new
field `trigger`, the value of which can be either "autoSave" or
"manualSave". The object used to have only one field (`data`).