unreliable focus detection could lead to changes bypassing the undo stack
Background: When a webapp is focussed, it receives both keyboard events
and mouse events. When a parent frame or another window is focussed, the
webapp receives mouse events, but not keyboard events.

In such a situation, Shift+MouseMove can be processed and change the
annotation data, but the Keyup(ShiftKey) that would add the change to
the undo stack is never received. So a user can make changes but the app
still indicates “data unchanged.”

In order to prevent this, we have the focus detection in place. Its aim
is to ignore MouseEvents when the app does not have focus.

However, the focus detection was unreliable. At startup, it assumed to
have focus and only changed the state when receiving focus events or
blur events. While this is fine most of the time, the app is not
actually guaranteed to have focus at startup. In particular, when loaded
via an iframe, the parent frame is somewhat likely to retain focus.

I replaced the event-driven (focus and blur events) focus detection by
the simple browser API document.hasFocus().