Minor fix for wav parser
A WAV header contains a BitsPerSample field, which was (apparently due
to a typo: 12 instead of 22) searched at the wrong index in the header.
This led to a wrong calculation of the length parameter to the
OfflineAudioContext() constructor. However, this appears to have been
inconsequential.

Also, the parser started looking for the data chunk of WAV files 10
bytes to early (x + 14 instead of x + 24). This could only have been a
problem if the string 'data' head been found in these 10 bytes, which is
basically impossible (the bytes encode numbers and not characters).